fix: build with Xcode 10 may fail for missing plugins...xcconfig file
In the iOS Project template we have include for plugins-debug.xcconfig (or plugins-release.xcconfig when building in release mode). CLI should produce such file, but it does not do it when the project does not have build.xcconfig in its App_Resources and neither of the plugins have such file. With previous Xcode versions, this case was leading to warnings in the output, however they are errors now and the build fails.
Fix this by adding an empty file in such case.